// src/app/Models/Customer.ts
export class Customer{
  id: number;
  shortName: string;
  longName: string;
  phone: string;
  idCardNumber: string;
  postcode: string;
  city: string;
  address: string;
  country: string;
  discount: number;
  email: string;
  nip: string;
  regon: string;
  description: string;
  bankAccountNumber: string;
}